pyproject.toml
[project]
name = "12-adboost-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.24.0",
    "openai-agents>=0.10.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]
